pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rest_api_ingestion"
version = "0.1.0"
description = "A REST API ingestion system"
authors = [{ name = "Pranav P Nair", email = "297027@ust.com" }]
dependencies = [
    "requests",
    "PyYAML",
    "python-dotenv",
    "Flask",
    "psycopg2",
    "pytest",
    "schedule"
]

[tool.pytest.ini_options]
pythonpath = "."
